backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
ipython==8.18.1
requests==2.31.0
openai>=1.54.0
httpx>=0.27.0
gunicorn==21.2.0